pagerduty_client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mai-util"
version = "0.1.0"
description = "MAI Shared Utility Library"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [
    {name = "MAI Team", email = "dev@mai.co"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "google-cloud-secret-manager>=2.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=5.0.0",
    "black>=25.1.0",
    "ruff>=0.13.1",
    "mypy>=1.17.1",
]

[tool.setuptools]
packages = ["mai_util", "mai_util.pagerduty_client"]

[tool.setuptools.package-data]
mai_util = ["py.typed"]

[tool.black]
line-length = 120
target-version = ['py311', 'py312']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disable_error_code = "union-attr"

[tool.pytest.ini_options]
testpaths = ["tests"]
